+++ src/oeis/oeis_program.cpp
@@ -90,6 +90,7 @@ bool OeisProgram::unfold(Program &p) {
   if (ProgramUtil::hasIndirectOperand(p)) {
     return false;
   }
+  // TODO unfold prg instructions, too
   // find first seq operation
   int64_t seq_index = -1;
   for (size_t i = 0; i < p.ops.size(); i++) {
@@ -117,30 +118,16 @@ bool OeisProgram::unfold(Program &p) {
     }
   }
   // find cells that are read and uninitialized
-  std::set<Operand> initialized, uninitialized;
-  initialized.insert(Operand(Operand::Type::DIRECT, Program::INPUT_CELL));
-  for (const auto &op : p2.ops) {
-    auto meta = Operation::Metadata::get(op.type);
-    if (meta.num_operands > 0 && op.target.type == Operand::Type::DIRECT) {
-      if (meta.is_reading_target &&
-          initialized.find(op.target) == initialized.end()) {
-        uninitialized.insert(op.target);
-      }
-      if (meta.is_writing_target) {
-        initialized.insert(op.target);
-      }
-    }
-    if (meta.num_operands > 1 && op.source.type == Operand::Type::DIRECT) {
-      if (initialized.find(op.source) == initialized.end()) {
-        uninitialized.insert(op.source);
-      }
-    }
-  }
+  std::set<int64_t> initialized, uninitialized;
+  initialized.insert(Program::INPUT_CELL);
+  ProgramUtil::getUsedUninitializedCells(p2, initialized, uninitialized);
+
   // initialize cells that are read and were uninitialized
   for (auto cell : uninitialized) {
-    p2.ops.insert(p2.ops.begin(),
-                  Operation(Operation::Type::MOV, cell,
-                            Operand(Operand::Type::CONSTANT, 0)));
+    p2.ops.insert(
+        p2.ops.begin(),
+        Operation(Operation::Type::MOV, Operand(Operand::Type::CONSTANT, cell),
+                  Operand(Operand::Type::CONSTANT, 0)));
   }
   // shift used operands
   int64_t target = p.ops[seq_index].target.value.asInt();