+++ CHANGELOG.md
@@ -2,6 +2,11 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ## [Unreleased]
 
+### Enhancements
+
+* Internal `replace` command
+* Enhanced `unfold` command
+
 ## v24.9.2
 
 ## Featues